feat: refactor icon imports for consistency and clarity
