+++ app/components/icons/assessment/index.ts
@@ -1,0 +1,3 @@
+export { default as DownloadIcon } from "./DownloadIcon";
+export { default as PlayIcon } from "./PlayIcon";
+export { ToggleOnIcon, ToggleOffIcon } from "./ToggleThumbIcons";
